感染状況・医療提供体制の分析: Rename components/MonitoringCommentCardImageSwipe.vue to components/index/SiteTopUpper/MonitoringComment/ImageSwipe.vue
